requirements.txt
pandas
pytest
streamlit
streamlit-lottie
requests
python-dotenv

